PROMISE-BASED REFACTOR
Hello folks! I have refactored parts of our flow control to work based on the concept of a "promise",
which is an abstract thing to be done after which another thing to be done can happen, but which is not
necessarily a sequence of code executing on a single thread. For example, a promise can represent a task
completed on another thread or the act of displaying a dialog (which is done when a user presses a button)
You can chain promises together to create a flow control, and also intermingle runnables and promises
to create a flow control based on doing things that require asynchronity and running some glue code in
between.

For example, to do a and then b, you'd say

myFancyClass.doA().then(myOtherFancyClass.doB()).go()

the .go() is important because promises don't run until their .go() method is called. Note that .doA() and .doB() are assumed to be special methods
that return promises, meaning that they don't do anything until their promises are executed.

to do a, then run an arbitrary method, then do B, you would write

myFancyClass.doA().then(new Runnable() {
	public void run() {
		runRandomMethod();
	}
}.then(myOtherFancyClass.doB()).do();

Hopefully, we can use this to write a nice tutorial. I will soon refactor other parts of our flow control into promises
so that we can more easily grasp WTF is happening when our app is run.